CCodec: fix underflow issue on handleImageData
the logic is assumed that gralloc does assume a contiguous mapping
at GraphicView2MediaImageConverter() in Codec2Buffer.
if it doesn't, underflow could happen because
type of variable is unsigned.

Bug: 168757280

+++ media/codec2/sfplugin/CCodecBuffers.cpp
@@ -91,7 +91,7 @@ void CCodecBuffers::handleImageData(const sp<Codec2Buffer> &buffer) {
             newFormat->setInt32(KEY_STRIDE, stride);
             ALOGD("[%s] updating stride = %d", mName, stride);
             if (img->mNumPlanes > 1 && stride > 0) {
-                int32_t vstride = (img->mPlane[1].mOffset - img->mPlane[0].mOffset) / stride;
+                int32_t vstride = (int32_t)(((int64_t)img->mPlane[1].mOffset - (int64_t)img->mPlane[0].mOffset) / stride);
                 newFormat->setInt32(KEY_SLICE_HEIGHT, vstride);
                 ALOGD("[%s] updating vstride = %d", mName, vstride);
             }
